Add changed files
